update id type

+++ db/migration/000001_init_schema.up.sql
@@ -1,20 +1,20 @@
 CREATE TABLE "accounts" (
-  "id" bigint PRIMARY KEY,
+  "id" bigserial PRIMARY KEY,
   "owner" varchar NOT NULL,
-  "balance" int NOT NULL,
+  "balance" bigint NOT NULL,
   "currency" varchar NOT NULL,
   "created_at" timestamp NOT NULL DEFAULT 'now()'
 );
 
 CREATE TABLE "entries" (
-  "id" bigint PRIMARY KEY,
-  "account_id" int NOT NULL,
+  "id" bigserial PRIMARY KEY,
+  "account_id" bigint NOT NULL,
   "amount" bigint NOT NULL,
   "created_at" timestamp NOT NULL DEFAULT 'now()'
 );
 
 CREATE TABLE "transfers" (
-  "id" bigint PRIMARY KEY,
+  "id" bigserial PRIMARY KEY,
   "from_account_id" bigint NOT NULL,
   "to_account_id" bigint NOT NULL,
   "amount" bigint NOT NULL,
